LPS-195582 Inline DirectRequestDispatcherFactoryImpl logic into Util
